Refactored the code to include a Page class
It seemed like a page class handling a lot of the work of finding links and
videos on it self seemed more appropriate. So refactored the code to include
such a class. Which cleaned up the DomainScraper quite a bit.

At the moment it still only checks to see if there is a iframe or not.